Add idiomatic EmberScript generators
A response to PR #206, these are the "idiomatic" CoffeeScript
generators, re-branded as EmberScript. They are essentially the same
changes but instead of renaming the original CoffeeScript files, we're
adding *.js.em files to the repo to separate Ember-infused CoffeeScript
with the more reliable approach as written initially in the CoffeeScript
generators.